trusted and clobber args added

+++ tests/test_core.py
@@ -51,13 +51,17 @@ def test_arg_check():
     with pytest.raises(ValueError, match='func must be callable'):
         xp.persist_ds('not a function')
 
+
 def test_make_cache_dir():
     old = xp.settings['cache_dir'] = os.path.join(here, 'cached_data')
     new = os.path.join(here, 'tmp_cached_data')
 
-    shutil.rmtree(new)
+    if os.path.exists(new):
+        shutil.rmtree(new)
     xp.settings['cache_dir'] = new
+
     ds = xp.persist_ds(func, name='test-dset')(10)
+
     assert os.path.exists(new)
 
     shutil.rmtree(new)
@@ -67,7 +71,28 @@ def test_make_cache_dir():
 
 def test_xpersist_noname():
     ds = xp.persist_ds(func)(10)
+    file, action = xp.persisted_Dataset._actions.popitem()
+    assert action == 'create_cache'
 
+
+def test_clobber():
+    ds = xp.persist_ds(func, name='test-dset')(10)
+    file, action = xp.persisted_Dataset._actions.popitem()
+    assert action == 'read_cache_verified'
+
+    ds = xp.persist_ds(func, name='test-dset', clobber=True)(11)
+    file, action = xp.persisted_Dataset._actions.popitem()
+    assert action == 'overwrite_cache'
+
+
+def test_trusted():
+    ds = xp.persist_ds(func, name='test-dset')(10)
+    file, action = xp.persisted_Dataset._actions.popitem()
+    assert action == 'read_cache_verified'
+
+    ds = xp.persist_ds(func, name='test-dset', trust_cache=True)(11)
+    file, action = xp.persisted_Dataset._actions.popitem()
+    assert action == 'read_cache_trusted'
 
 def test_validate_dset():
     dsp = xp.persist_ds(func, name='test-dset')(10)
